【submodule】Add glog use demo

+++ build.sh
@@ -5,12 +5,16 @@
 # Created Time: 2024-11-22 17:20:45
 #########################################################################
 #!/bin/bash
-rm -rf build
+CLEAN="noclean"
+[ -n "${1}" ] && CLEAN="${1}"
+if [ -n "${CLEAN}" ] && [ "${CLEAN}" = "clean" ]; then
+  rm -rf build
+fi
 mkdir build
 cd build
 cmake ../
 cmake --build .
-#./src/cmakehello
-#ldd ./src/cmakehello
 cd ../
-
+ldd ./build/src/main/gloghello
+./build/src/main/gloghello 100
+ls -al /tmp/logs/
